requirements.txt
torch>=2.2
torchvision>=0.17
timm>=1.0
gymnasium>=1.0
numpy>=1.26
matplotlib>=3.8
pyyaml>=6.0
tqdm>=4.66
pygame>=2.5
